requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
boto3==1.26.153
botocore==1.29.153
certifi==2023.5.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
contourpy==1.1.0
cycler==0.11.0
fonttools==4.40.0
frozenlist==1.3.3
gym==0.26.2
gym-notices==0.0.8
idna==3.4
ikpy==3.1
imageio==2.31.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
jmespath>=0.10.0
kiwisolver==1.4.4
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.0.4
numpy==1.24.3
openai==1.30.1
opencv-python==4.7.0.72
overrides==7.3.1
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
psutil==5.9.5
pyastar2d==1.0.6
pydub==0.25.1
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
requests>=2.28.2
s3transfer==0.6.1
scipy==1.10.1
screeninfo==0.8.1
seaborn==0.12.2
six==1.16.0
sympy==1.12
tdw==1.12.27.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.16
yarl==1.9.2
zipp==3.15.0
gdown==5.2.0